Added initial implementation of the brick wall system.  Bullets are working with some amount of jank (seems like fire rate slows over time - investigating).  Walls are spawning, visible, and destructible.  Collisions are working.
The art for the walls is placeholder, and needs to be updated.  The walls aren't centered on the pipes, and they're stretched vertically by a bit (need to get them centered and make them a bit wider).  May put together a simple particle system for the walls before I call this project completed.

I removed the vsync stuff.  Don't need to worry about uncapped framerate for a game this simple, the default ~62 FPS should be plenty.